+++ examples/merge_typed_functions.js
@@ -1,6 +1,6 @@
 var typed = require('../typed-function');
 
-// create a number of typed functions
+// create a couple of typed functions
 var fn1 = typed('number', function (a) {
   return a * a;
 });
@@ -13,5 +13,5 @@ var fn2 = typed('string', function (a) {
 var fn3 = typed(fn1, fn2);
 
 // use merged function
-console.log(fn3(2)); // outputs 4
-console.log(fn3('3')); // outputs 9
+console.log(fn3(2));    // outputs 4
+console.log(fn3('3'));  // outputs 9