+++ HISTORY.md
@@ -2,6 +2,7 @@
 
 ## not yet released, version 0.5.0
 
+- Implemented support for merging typed functions.
 - Fixed a bug where a regular argument was not matched when there was a
   signature with variable arguments too.